refactor(code): keep one public dual and syndrome operation ID
`code.dual_code.compute` and `code.syndrome.compute` routed to the
identical request type, result type, and callable as the published
`code.linear.dual.compute` and `code.linear.syndrome.compute`,
leaving two public IDs for one postcondition and letting dual/syndrome
discovery return indistinguishable alternatives. Remove the duplicate
code_theory declarations and their admissions so each postcondition
keeps its single canonical code_linear ID.

Relocate the retained regression evidence to its owning lane: Hamming
[7,4] dual and ternary-syndrome known answers join the code_linear
math tests, and the serialized dual-into-syndrome dispatch chain runs
against the canonical IDs under tests/dispatch. Removing the old test
file also clears the jacobian.catalog/jacobian.dispatch imports from
tests/math that failed the tooling collection-ownership guard.
